refactored home page to reflect same scrolling behavior
