+++ client/src/components/Values/index.jsx
@@ -5,20 +5,7 @@ import { Link } from 'react-router-dom'
 import { Container, Title, Text, Button } from '@mantine/core'
 import classes from '@/assets/css/HeroImageRight.module.css'
 
-const Values = () => {
-  const [scrollY, setScrollY] = useState(0)
-
-  const handleScroll = () => {
-    setScrollY(window.scrollY)
-  }
-
-  useEffect(() => {
-    window.addEventListener('scroll', handleScroll)
-    return () => {
-      window.removeEventListener('scroll', handleScroll)
-    }
-  }, [])
-
+const Values = ({ scrollY }) => {
   return (
     <div
       className={classes.root}
